interim, restructuring + db
